it is usually played offline though
So your task is to guess 3 digits from 0 to 9. Digits do not repeat. If you guess a digit and its location, the game will tell you "1 beef" or "2 beef'', if you guess a digit but not its location, the game will say "1-3 chicken". If you didn't guess anything, you will get a 'veg' message. Try to guess a number in the least amount of turns.